fix: shared hardware elements

+++ nixos/modules/core/hardware/config.nix
@@ -1,20 +1,26 @@
 {
   host,
   pkgs,
+  system,
   lib,
   ...
 }:
 {
-  nixpkgs.overlays = [
-    (pkgs: prev: {
-      linux-firmware = prev.linux-firmware.overrideAttrs (o: {
-        postInstall = ''
-          rm -rf "$out"/lib/firmware/{netronome,qcom,mellanox,mrvl,ath11k,ath10k,libertas,nvidia,liquidio,cxgb4,ti-connectivity,qed}
-          find -L "$out" -type l -delete
-        '';
-      });
-    })
-  ];
+  nixpkgs.overlays = (
+    if system != "aarch64-linux" then
+      [
+        (pkgs: prev: {
+          linux-firmware = prev.linux-firmware.overrideAttrs (o: {
+            postInstall = ''
+              rm -rf "$out"/lib/firmware/{netronome,qcom,mellanox,mrvl,ath11k,ath10k,libertas,nvidia,liquidio,cxgb4,ti-connectivity,qed}
+              find -L "$out" -type l -delete
+            '';
+          });
+        })
+      ]
+    else
+      [ ]
+  );
 
   hardware = {
     enableAllFirmware = lib.mkForce false;
